# REF/PERF: ArrowStringArray.__setitem__