@lukemanley 2022-03-17
> Does this do anything to help with #45419?

No, it does not. If it is decided to make these immutable, then this isn't needed.