@jreback 2022-03-18
cc @jbrockmendel @mroeschke for comments on that

prob ok to add directly to ArrowExtensionArray i think